requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
anyio==4.2.0
astroid==3.0.2
attrs==23.2.0
autopep8==2.0.4
avro==1.11.3
avro-schema==0.3.3
azure-core==1.29.6
azure-eventhub==5.11.5
azure-identity==1.15.0
azure-storage-blob==12.19.0
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.7.0
bs4==0.0.1
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.7
cssselect==1.2.0
decorator==5.1.1
dicttoxml==1.7.16
dill==0.3.7
docutils==0.20.1
Flask==3.0.0
Flask-WTF==1.2.1
frozenlist==1.4.1
future==0.18.3
geocoder==1.38.1
geoip2==4.8.0
h11==0.14.0
html5lib==1.1
idna==3.6
importlib-metadata==7.0.1
ip2geotools==0.1.6
IP2Location==8.10.1
isodate==0.6.1
isort==5.13.2
itsdangerous==2.1.2
jaraco.classes==3.3.0
Jinja2==3.1.3
keyring==24.3.0
lazy-object-proxy==1.10.0
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
maxminddb==2.5.2
mccabe==0.7.0
mdurl==0.1.2
mechanize==0.4.9
more-itertools==10.2.0
msal==1.26.0
msal-extensions==1.0.0
multidict==6.0.4
mysql-connector-python==8.2.0
nh3==0.2.15
outcome==1.3.0.post0
packaging==23.2
pip-review==1.3.0
pkginfo==1.9.6
platformdirs==4.1.0
portalocker==2.8.2
protobuf==4.21.12
pycodestyle==2.11.1
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
pylint==3.0.3
pyparsing==3.1.1
pyquery==2.0.0
PySocks==1.7.1
python-geoip==1.2
ratelim==0.1.6
readme-renderer==42.0
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.0
selenium==4.16.0
setuptools==69.0.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
tomlkit==0.12.3
tqdm==4.66.1
trio==0.24.0
trio-websocket==0.11.1
twine==4.0.2
typed_ast==1.5.5
typing==3.7.4.3
typing_extensions==4.9.0
urllib3==2.1.0
webencodings==0.5.1
Werkzeug==3.0.1
wrapt==1.16.0
wsproto==1.2.0
WTForms==3.1.2
yarl==1.9.4
zipp==3.17.0
